change id long to int

+++ src/main/java/com/ArthurGrand/admin/tenants/controller/TenantController.java
@@ -31,7 +31,7 @@ public class TenantController {
     }
 
     @PutMapping("/{tenantId}/activate")
-    public ResponseEntity<?> activateTenant(@PathVariable Long tenantId) {
+    public ResponseEntity<?> activateTenant(@PathVariable int tenantId) {
         try {
             TenantResponseDto response = tenantService.activateTenant(tenantId);
             return ResponseEntity.ok()
