+++ src/main/java/com/ArthurGrand/admin/tenants/repository/TenantProfileRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface TenantProfileRepository extends JpaRepository<TenantProfile,Long> {
-    Optional<TenantProfile> findByTenantId(Long tenantId);
+public interface TenantProfileRepository extends JpaRepository<TenantProfile,Integer> {
+    Optional<TenantProfile> findByTenantId(int tenantId);
 }
